+++ package.json
@@ -14,11 +14,13 @@
 	"dependencies": {
 		"@vueup/vue-quill": "^1.2.0",
 		"axios": "^1.7.7",
+		"chart.js": "^4.1.1",
 		"firebase": "^10.14.0",
 		"js-cookie": "^3.0.5",
 		"pinia": "^2.2.4",
 		"quill": "^2.0.2",
 		"vue": "^3.4.29",
+		"vue-chartjs": "^5.3.1",
 		"vue-router": "^4.4.5"
 	},
 	"devDependencies": {
